mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.4",
    "mysql-connector-python>=9.3.0",
]
